updated route to have check for if lown exists

+++ app/api/loan_routes.py
@@ -7,20 +7,24 @@ loan_routes = Blueprint("loan", __name__)
 @loan_routes.route("/<int:id>", methods=["GET", "PUT", "DELETE"])
 def loans(id):
     loan = Loan.query.get(id)
-    if request.method == "PUT":
-        payload = json.loads(request.data)
-        loan.amount = payload["amount"] if "amount" in payload else loan.amount
-        loan.interest_rate = payload["interest_rate"] if "interest_rate" in payload else loan.interest_rate
-        loan.loan_length = payload["loan_length"] if "loan_length" in payload else loan.loan_length
-        loan.monthly_payment = payload["monthly_payment"] if "monthly_payment" in payload else loan.monthly_payment
-        db.session.commit()
-        return {"message": "Loan updated", 
-                "loan": loan.to_dict()}
-    elif request.method == "DELETE":
-        db.session.delete(loan)
-        db.session.commit()
-        return {"message": "Delete Successful."}
-    return {loan.id: loan.to_dict()}
+    
+    if loan:
+        if request.method == "PUT":
+            payload = json.loads(request.data)
+            loan.amount = payload["amount"] if "amount" in payload else loan.amount
+            loan.interest_rate = payload["interest_rate"] if "interest_rate" in payload else loan.interest_rate
+            loan.loan_length = payload["loan_length"] if "loan_length" in payload else loan.loan_length
+            loan.monthly_payment = payload["monthly_payment"] if "monthly_payment" in payload else loan.monthly_payment
+            db.session.commit()
+            return {"message": "Loan updated", 
+                    "loan": loan.to_dict()}
+        elif request.method == "DELETE":
+            db.session.delete(loan)
+            db.session.commit()
+            return {"message": "Delete Successful."}
+        return {loan.id: loan.to_dict()}
+    else:
+        return{"message": "Loan not found"}
 
 
 @loan_routes.route("/", methods=["GET", "POST"])
